Use container format specified in transfer project profile

+++ ESSArch_TA/ip/views.py
@@ -293,9 +293,15 @@ class InformationPackageViewSet(viewsets.ModelViewSet):
         srcdir = Path.objects.get(entity="path_ingest_prepare").value
         dstdir = Path.objects.get(entity="path_gate_reception").value
 
-        src = os.path.join(srcdir, "%s.tar" % pk)
-        dst = os.path.join(dstdir, "%s.tar" % pk)
-        shutil.copy(src, dst)
+        if os.path.isfile(os.path.join(srcdir, "%s.tar" % pk)):
+            src = os.path.join(srcdir, "%s.tar" % pk)
+            dst = os.path.join(dstdir, "%s.tar" % pk)
+            shutil.copy(src, dst)
+
+        if os.path.isfile(os.path.join(srcdir, "%s.zip" % pk)):
+            src = os.path.join(srcdir, "%s.zip" % pk)
+            dst = os.path.join(dstdir, "%s.zip" % pk)
+            shutil.copy(src, dst)
 
         src = os.path.join(srcdir, "%s.xml" % pk)
         dst = os.path.join(dstdir, "%s.xml" % pk)
